requirements.txt
black==22.3.0
flake8==4.0.1
numpy==1.22.4
isort==5.10.1
pandas==1.4.2
pre-commit==2.19.0
pytest==7.1.2
python-dotenv==0.20.0
google-cloud-logging>=3.0.0
python-dateutil==2.8.2
